Module madara::import_queue

Helpers for building the block import queue and pipeline.

Block import queue is an implementation of the chain responsibility pattern, where the very last handler is the client and each preceeding handler wraps the current one forming an extra layer.

Read more about the block import pipeline:

https://docs.substrate.io/learn/transaction-lifecycle/#block-authoring-and-block-imports

https://doc.deepernetwork.org/v3/advanced/block-import/

https://substrate.stackexchange.com/search?q=block+import

In order to avoid confusion: - Block import is a trait that all handlers in the queue/pipeline have to implement; - Import queue is a struct that sc_service::build_network accepts as a configuration parameter; - Import pipeline is a helper struct that encapsulates a BlockImport used to import own (authored) blocks plus optionally a link (hook) to Grandpa consensus.

Import queue is used to import external blocks (created by other nodes), so typically it would look like: Aura: ImportQueueVerifier -> [Grandpa: BlockImport -> Client: BlockImport]

For own (authored) blocks the chain looks similar (although a bit different): Aura: BlockAuthoringTask -> [Grandpa: BlockImport -> Client: BlockImport]

Read more about consensus: * https://docs.substrate.io/learn/consensus/ * https://substrate.stackexchange.com/questions/5918/how-can-i-make-my-node-generate-blocks-only-when-they-receive-transactions

Structs

BlockImportPipeline

Block import pipeline is a helper struct encapsulating the authored (own) block import and optionally a link (hook) to the Grandpa block import.

Constants

GRANDPA_JUSTIFICATION_PERIOD

The minimum period of blocks on which justifications will be imported and generated.

Functions

build_aura_import_queue 🔒

Build the import queue for default sealing given the block import.

build_aura_queue_grandpa_pipeline

Build a block import queue & pipeline for default sealing.

build_manual_seal_queue_pipeline

Build a block import queue & pipeline for manual/instant sealing.

Type Definitions

GrandpaLinkT 🔒

Grandpa link (hook) type.